Break `ItemPicker` into view and container
